Add room. Update components version. Add prefilled db with cities data.

+++ app/src/main/java/com/ringov/yamblzweather/model/networking/data/Wind.java
@@ -7,16 +7,16 @@ public class Wind {
 
     @SerializedName("speed")
     @Expose
-    private Integer speed;
+    private Double speed;
     @SerializedName("deg")
     @Expose
     private Integer deg;
 
-    public Integer getSpeed() {
+    public Double getSpeed() {
         return speed;
     }
 
-    public void setSpeed(Integer speed) {
+    public void setSpeed(Double speed) {
         this.speed = speed;
     }
